Improve shell script and instructions
Insufficient quoting in shell script; having "rm -rf" in a script
deleting something not created with `mktemp` gives me the willies.
Especially if I make quoting errors... But we can do it with git, that
is a lot safer.

Passing `-Wall` to the `clash` invocation is a no-op since it is not
compiling anything. Cabal already compiled the design by the time the
`clash` binary accesses it, and the `.cabal` file already specifies
`-Wall -Wcompat`. This means any warnings have already been emitted.